+++ src/main/java/com/assesment/inventory_management/model/Product.java
@@ -18,7 +18,7 @@ public class Product {
     private String name;
 
     @Column(nullable = false)
-    private Integer price;
+    private BigDecimal price;
 
     private Integer quantity;
 
@@ -36,7 +36,7 @@ public class Product {
 
     public Product() {    }
 
-    public Product(String name, Integer price, Integer quantity, Category category) {
+    public Product(String name, BigDecimal price, Integer quantity, Category category) {
         this.name = name;
         this.price = price;
         this.quantity = quantity;
@@ -59,11 +59,11 @@ public class Product {
         this.name = name;
     }
 
-    public Integer getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
